+++ templates/app/assets/images/uswds.js
@@ -1,5 +1,6 @@
 // Glue to find USWDS images with the `image_tag` helper
 
-//= link uswds/dist/img/us_flag_small.png
-//= link uswds/dist/img/icon-dot-gov.svg
-//= link uswds/dist/img/icon-https.svg
+//= link @uswds/uswds/dist/img/us_flag_small.png
+//= link @uswds/uswds/dist/img/icon-dot-gov.svg
+//= link @uswds/uswds/dist/img/icon-https.svg
+//= link @uswds/uswds/dist/img/usa-icons/close.svg